Initial version of the barcode-extraction code.
This extraction method is intended to be very precise andwork in
high-noise images (text, other barcodes, etc). It has three stages:

 * First, the AnchoredScan is used to located candidates for the
   the edge guards. So far, I've just averaged the pixels to the
   immediate left and right of the anchor (the black bar) to test
   for edge guard-ness. This generally finds the edge guards very
   well, but also detects spurious guards.
 * To get rid of these false positives, the Hough transform (for
   lines) is run over the detected points. The idea is that all
   points on the edge guard--and this should be about 70% of all
   points--will be collinear. The others will be dropped, and the
   line of best fit saved.
 * From there, the GuardTracer is given a point on the edge guard
   and traces the guard (using the line from HoughLines) to the
   corner. This works well on well-aligned barcodes, but rotation
   throws it off.

After the corners are located, the barcode can be undistorted using
ImageMagick Distort Perspective, extracted, or whatever needs to be
done.

There's a lot of work to be done. The GuardTracer does not work for
vertical or horizontal lines; both scanning and hough could be
optimized; and the whole process could be improved by only scanning
areas of high variation; and the errors are very unhelpful.