This commit introduces the "adaptive" anchor scanner.
Unlike the previous version, it requires very few image-specific
parameters, while achieving higher success rates.

The AdaptiveAnchorScan works on more or less the same principle as
the former GuardScan, except that the black-white threshold is
calculated locally within the image: for every NxM block within the
image, a block is run that returns an AdaptiveThreshold object that
gives the "gray" value (useful for determining intensity shifts) and
a deviation value.

The deviation value is used to make the scanner more stable---that
is, when it's scanning black, it accepts some brighter values as
black, and when scanning white, it accepts some darker intensities.

There's also a blurring parameter, that I hope will improve the
algorithm in noisy images. To save time, pixels are only blurred on
failure.

A major difference between the new scanner and the former is that
the new version returns the best guard match on every line. The
scoring, like the thresholding, is performed in a block passed to
the AdaptiveAnchorScan object. This does result in many false
positives, but they are very scattered and the collinearity detector
should have no trouble with them.

There's also a testing_code.rb file that contains ~200 lines of
playground material. This will not be saved, I've only included it
here in case it becomes useful later.