Added non-maximum suppression to Hough line detection
This is fairly simple: for every value in the accumulator, determine
if there is any other value within a given window size such that the
first point is less than to the second. If so, that first point
isn't a maximum. If the points are equal, the first point is set to
0, in order to avoid future competition. In most cases, once the
competition is thus eliminated, the second point will become the
eventual max. If the first point is larger than any other point in
its neighborhood, then it's considered a max.

This often results in small local maximums (values of 1 or 2 for
example can be maximums in a sparse region), so the collected
maximums are filtered by a threshold.

Also fixed a misspelling of Ludwing Otto Hesse's last name.