Removed a old file
The method used for edge guard detection used in this file is
several months old now, and has since been superseded. It was called
precise because, unlike the new method, attempted to perform pixel-
perfect localization. In this it was only partially successful, as
it required very clean, well-oriented images to work well.