Implemented code to roughly located a barcode
The building blocks are in place, at least. A fair bit of work
remains, especially around the guard detection qualification area.

Given a set of points on the edge guard line, RoughLocalization
uses the regional densities of those points to estimate where the
edge guard begins and ends. From there, it fuzzes those estimates
in order to be sure of including all the barcode.

The fuzzing code could be improved, but it works well for now.

This method is called "rough" because it doesn't attempt pixel-
perfect detection at all. Nevertheless, the detection will not be
more than about 60% the given window size from the true, probably.
This generally comes to about 6-12 pixels, which is pretty precise.