Implemented Ramer-Douglas-Peucker polyline simplification
Had to use some different data structures than anticipated --- the
advantages of the GList weren't as useful as hoped, so switched
over to a GPtrArray for the region collection (the RDRegion
struct's are shared between the hash and the array, so pointers are
necessary) and a GArray for the contour point collection (avoids
explicit copying that would be necessary for GPtrArray's).

Also wrote some tests.