Extend aspect settings and revise barcode scoring
Changed guard_aspect setting of Localization::Guards to be a range
(values should generally be between 3 and 30), and added the
barcode_aspect setting, which was previously tested by
Localization::Guards#positioned_well?.

Also improved the scoring algorithm to account for rectangle area
similarity. It's still a bit aggressive, we'll probably have to
continue to tweak it.