Optimized darray_msort slightly and investigated darray_qsort
I'd originally hoped to write darray_qsort, which could be faster
and would use less memory. The idea was to use qsort_r where
possible, and fall back to a custom implementation otherwise.
Unfortunately, darray_qsort turned out to be slow and ugly, and
like all quicksorts, unstable (mergesorts are stable). It did
indeed use significantly less memory than darray_msort, but I didn't
think it was worth it.

Here are the benchmarks (qsort included for baseline, it is
too different to be used in practice):

  Sorting 10000000 random integers
  darray_msort: 1.62 seconds
  darray_qsort: 1.39 seconds (custom implementation)
  darray_qsort: 1.45 seconds (GNU qsort_r implementation)
  qsort: 1.29 seconds

  Sorting 10000000 pre-sorted integers
  darray_msort: 0.51 seconds
  darray_qsort: 0.39 seconds (custom implementation)
  darray_qsort: 0.48 seconds (GNU qsort_r implementation)
  qsort: 0.38 seconds

  Sorting 10000000 random integers between 0 and 64
  darray_msort: 1.02 seconds
  darray_qsort: **took too long** (custom implementation)
  darray_qsort: 1.05 seconds (GNU qsort_r implementation)
  qsort: 0.88 seconds

  Sorting 10000000 random strings
  darray_msort: 4.74 seconds
  darray_qsort: 5.49 seconds (custom implementation)
  darray_qsort: 4.26 seconds (GNU qsort_r implementation)
  qsort: 4.10 seconds

Once I fix the duplicate element handling (as in "random integers
between 0 and 64"), though, darray_qsort will be competitive.